[sw/rom_ext] Rename rom_ext folder.
sw/device/silicon_creator/rom_exts to sw/device/silicon_creator/rom_ext
to avoid ambiguity.
