+++ sw/device/_index.md
@@ -10,7 +10,7 @@ Different images are used for different boot stages.
 The Reference Firmware Images are, in boot order:
 
 1.  The [Mask ROM]({{< relref "sw/device/silicon_creator/mask_rom/docs" >}}) (in `sw/device/mask_rom`), executed at chip reset;
-2.  The ROM_EXT (in `sw/device/silicon_creator/rom_exts`), the second stage Silicon Creator code, executed from flash; and
+2.  The ROM_EXT (in `sw/device/silicon_creator/rom_ext`), the second stage Silicon Creator code, executed from flash; and
 
 ### Testing-only Images
