+++ sw/device/silicon_creator/rom_ext/docs/manifest.md
