+++ sw/device/silicon_creator/rom_ext/rom_ext_boot_policy_unittest.cc
@@ -2,11 +2,11 @@
 // Licensed under the Apache License, Version 2.0, see LICENSE for details.
 // SPDX-License-Identifier: Apache-2.0
 
-#include "sw/device/silicon_creator/rom_exts/rom_ext_boot_policy.h"
+#include "sw/device/silicon_creator/rom_ext/rom_ext_boot_policy.h"
 
 #include "gtest/gtest.h"
 #include "sw/device/silicon_creator/lib/mock_manifest.h"
-#include "sw/device/silicon_creator/rom_exts/mock_rom_ext_boot_policy_ptrs.h"
+#include "sw/device/silicon_creator/rom_ext/mock_rom_ext_boot_policy_ptrs.h"
 #include "sw/device/silicon_creator/testing/mask_rom_test.h"
 
 namespace manifest_unittest {
