+++ sw/device/silicon_creator/rom_ext/rom_ext_common.ld
