+++ sw/device/silicon_creator/rom_ext/rom_ext_slot_a.ld
@@ -9,7 +9,7 @@
  *
  * The ROM_EXT is actually kept in flash, rather than ROM. While a ROM_EXT can
  * be loaded into either Slot A (the start of flash), or Slot B (the start of
- * the upper half of flash), this linker script only targets Slot B.
+ * the upper half of flash), this linker script only targets Slot A.
  */
 
 INCLUDE hw/top_earlgrey/sw/autogen/top_earlgrey_memory.ld
@@ -19,7 +19,7 @@ _stack_size = 0x2000;
 _stack_end = ORIGIN(ram_main) + LENGTH(ram_main);
 _stack_start = _stack_end - _stack_size;
 
-/* Slot B starts at the half-size mark of the eFlash. */
-_slot_start_address = ORIGIN(eflash) + (LENGTH(eflash) / 2);
+/* Slot A starts at the start of the eFlash. */
+_slot_start_address = ORIGIN(eflash);
 
-INCLUDE sw/device/silicon_creator/rom_exts/rom_ext_common.ld
+INCLUDE sw/device/silicon_creator/rom_ext/rom_ext_common.ld
